Implement log insertions
Program now allows developers to insert into an app feed. It stores each app feed's public key and its corresponding private key in a json file which it uses to append to the feed.

Also allowed the definition of assets and made the program distinguish between an APP entry and an asset entry and return feedback accordingly